[Modification] Now using movie structure from go-tmdb package instead of custom movie structure

+++ api/src/domain/user_favorites/user_favorites_dto.go
@@ -8,6 +8,6 @@ type UserFavoritesInterface interface {
 }
 
 type UserFavorites struct {
-	UserID    int64   `json:"user_id"`
-	MoviesIDs []int64 `json:"favorite_movies"`
+	UserID    int64 `json:"user_id"`
+	MoviesIDs []int `json:"favorite_movies"`
 }
